CityFormatter automatically adds commas before state names and capitalizes the first letter of city names
It doesn't have a constructor, just a static method called with
CityFormatter.format(string). This can be used when displaying
the city name and also when the user is searching for a city. (They
don't need to enter commas before the state anymore)

Added some test code to make sure its working.