ControlsFXTest now implements Initializable, this allows it so set binding when the app fires up.
bindAuto adds a string to the suggestions Set, then disposes old binding
and rebinds tfInput to the suggestions Set. This only happens if the
word isn't already in the set, and while doing so it adds the long,lat
to a HashMap to be used for later.

Sometimes multiple autocomplete boxes will still appear, however it is
doing it a lot less since using the dispose method.