Updated controller to use the combo box autocompletion method instead of the controlsfx one. Controlsfx was giving me too many problems to be able to use nicely.
Changed the input method to a combo box. Also had to use a bug fix
that Phil found on stack overflow about the space bar selecting the
current suggestion or sometimes closing the items window.
( https://stackoverflow.com/a/52061856 )

The autoComplete method searches using MapBox after at least 3 chars
are in the editor window, and only searches on every other char after
that. There is still an issue where the editor would sometimes clear the
currently typed text in the window after the user selects a suggestion,
so each time new results are fetched I replace the text with the same
text and position the caret at the end, this is not ideal but it is
working (Lines 123, 124). Finally if the user presses enter then the
go button's handler is fired, and the suggestions are hidden.